+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added event sources integration: API and on-prem consumer for SQS queues (event-driven ingestion) in `src/onprem/event_sources_consumer`
 - Added handling obsolete/removed rules in shards
 - Added the ability to get a resource report for a specific platform(K8S) job
+- Added the intermediate scan saves and reports (by region)
+- Added the ability to resume stuck jobs using intermediate data by `POST /jobs/{job_id}/resume` endpoint
+
+### Changed
+- Expanded the list of job error codes and messages
+- Changed terminal job status to `INTERRUPTED` instead of `FAILED`
 
 ### Fixed
 - Fixed on-prem API not passing query params (e.g. `customer_id`) for non-GET requests (DELETE, PUT, etc.)